# Improve concurrent optimization