--- a/mpicbg/src/main/java/mpicbg/models/TileConfiguration.java
+++ b/mpicbg/src/main/java/mpicbg/models/TileConfiguration.java
@@ -16,6 +16,8 @@
  */
 package mpicbg.models;
 
+import mpicbg.util.RealSum;
+
 import java.io.Serializable;
 import java.text.DecimalFormat;
 import java.text.DecimalFormatSymbols;
@@ -27,8 +29,9 @@ import java.util.List;
 import java.util.ListIterator;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
-
-import mpicbg.util.RealSum;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Future;
+import java.util.concurrent.ThreadPoolExecutor;
 
 
 /**
@@ -119,27 +122,43 @@ public class TileConfiguration implements Serializable
 	 */
 	protected void apply()
 	{
-//		final ArrayList< Thread > threads = new ArrayList< Thread >();
-//		for ( final Tile< ? > t : tiles )
-//		{
-//			final Thread thread = new Thread(
-//							new Runnable()
-//							{
-//								final public void run()
-//								{
-//									t.apply();
-//								}
-//							} );
-//			threads.add( thread );
-//			thread.start();
-//		}
-//		for ( final Thread thread : threads )
-//		{
-//			try { thread.join(); }
-//			catch ( InterruptedException e ){ e.printStackTrace(); }
-//		}
 		for ( final Tile< ? > t : tiles )
 			t.apply();
+	}
+
+	/**
+	 * Apply the model of each {@link Tile} to all its
+	 * {@link PointMatch PointMatches} using a given {@link ExecutorService}
+	 * and a given number of threads.
+	 */
+	protected void apply(final ThreadPoolExecutor executor) {
+		final List<Tile<?>> allTiles = new ArrayList<>(tiles);
+		final int nTiles = allTiles.size();
+		final int nThreads = executor.getActiveCount();
+		final int tilesPerThread = nTiles / nThreads + (nTiles % nThreads == 0 ? 0 : 1);
+		final List<Future<Void>> applyTasks = new ArrayList<>(nThreads);
+
+		for (int j = 0; j < nThreads; j++) {
+			final int start = j * tilesPerThread;
+			final int end = Math.min((j + 1) * tilesPerThread, nTiles);
+			applyTasks.add(executor.submit(() -> applyToRange(allTiles, start, end)));
+		}
+
+		for (final Future<Void> task : applyTasks) {
+			try {
+				task.get();
+			} catch (final InterruptedException | ExecutionException e) {
+				throw new RuntimeException(e);
+			}
+		}
+	}
+
+	private static Void applyToRange(final List<Tile<?>> tiles, final int start, final int end) {
+		for (int i = start; i < end; i++) {
+			final Tile<?> t = tiles.get(i);
+			t.apply();
+		}
+		return null;
 	}
 
 	/**