@minnerbe 2024-02-13
A follow up based on an offline discussion with @axtimwalde, who expressed interest in more validation in the form of real experiments.

I pointed some of our EM-alignment code at this PR and re-aligned a few stacks. One stack consists of 60k tiles with ~100 matches per tile pair. The models that are fitted are multi-layer interpolated models (i.e., an interpolation of affine, rigid, and translation).

The alignments were done with 8, 16, and 32 cores and compared against the original algorithm with 32 cores. For all runs, the alignment errors were basically identical, but runtime for the algorithm proposed in this PR was roughly 3, 5, and 7 times lower, respectively. In particular, the algorithm seems to consistently terminate and produce the same results as before the change. 

Is this experiment sufficient @axtimwalde, or do you want to see further results?