@acardona 2024-03-27
Only to add: THANKS!

On a 256-core computer, the optimizer went from spending over 8 hours to a mere 20 minutes, for a ~12,000 section series registered 6-adjacent with hundreds to a thousand SIFT features per pairwise section matching. This is AMAZING.